Move followage module to basic commands group (#954)
